Refactor/app bar (#116)
* feat(ui): Implement custom app bar

Adds a custom app bar and assets.

* refactor(ui): Refactor app bar and update UI

This commit refactors the app bar implementation by introducing
`SellioAppBar` and `HomeAppBar` components. It removes the
`CustomAppBar` and `sellio_back_app_bar` components. The commit also
includes UI updates to the `ForgetPasswordScreen`,
`SetNewPasswordScreen`, `ProductDetailsScreen`, and
`StoreDetailsScreen` components.

* refactor(ui): Refactor store data provider

Refactors the StoreDataProvider to a singleton, and
populates mock data for store details and products.