Implement Notifications Screen Navigation (#193)
* feat(navigation): Implement notifications screen route

Adds the route and navigation logic for the notifications screen.

- Defines a new `notifications` route in `AppRoutes`.
- Implements the navigation logic in the `AppNavigator` to handle navigating to the new screen.
- Configures the `GoRouter` to map the route to the `NotificationScreen`.
- Updates the home screen's navigation utility to navigate to the notifications screen when triggered.

* feat(navigation): Add notifications route

Adds the `pushNotifications` method to the `AppNavigator` interface to handle navigation to the notifications screen.

* Refactor: Update notification screen import path

Moves the `NotificationScreen` from `lib/presentation/screens/notification/notification_screen.dart` to `lib/presentation/screens/notification_screen.dart` and updates the import path in the route manager accordingly.